common: verbose: split profiler into base and xpu subclasses
